Google login succesfully implemented with signout options also

+++ lib/profile_page.dart
@@ -11,8 +11,11 @@ import 'package:car_rental_system/widgets/custom_circle_avatar.dart';
 import 'package:car_rental_system/widgets/custom_sized_box.dart';
 import 'package:car_rental_system/widgets/custom_text.dart';
 import 'package:car_rental_system/widgets/padding_for_all_pages.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:google_sign_in/google_sign_in.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class ProfilePage extends StatefulWidget {
@@ -35,10 +38,10 @@ class _ProfilePageState extends State<ProfilePage> {
   getUserIdAndInfo() async {
     userId = await GetUserInfo().getUserId();
     if (userId != null) {
-        Users? user1 = await GetUserInfo().getUserDetails(userId!);
-     setState(() {
-      user=user1;
-     });
+      Users? user1 = await GetUserInfo().getUserDetails(userId!);
+      setState(() {
+        user = user1;
+      });
     }
   }
 
@@ -132,10 +135,20 @@ class _ProfilePageState extends State<ProfilePage> {
                         onOkPressed: () async {
                           SharedPreferences prefs =
                               await SharedPreferences.getInstance();
-                          prefs.clear();
+                          bool logoutFromFirebaseGoogleLogin =
+                              await signOut(context: context);
+                          if (logoutFromFirebaseGoogleLogin) {
+                            prefs.clear();
+                            print("-------------------Logging out throught The firebase google account");
+                            RouteGenerator.navigateToPageWithoutStack(
+                                context, Routes.loginRoute,
+                                arguments: true);
+                          }else{
+                            print("#################### Logging out throught The register database account");
                           RouteGenerator.navigateToPageWithoutStack(
                               context, Routes.loginRoute,
                               arguments: true);
+                          }
                         });
                   },
                 ),
@@ -145,5 +158,24 @@ class _ProfilePageState extends State<ProfilePage> {
         ),
       ),
     );
+  }
+
+  Future<bool> signOut({required BuildContext context}) async {
+    final GoogleSignIn googleSignIn = GoogleSignIn();
+
+    try {
+      if (!kIsWeb) {
+        await googleSignIn.signOut();
+      }
+      await FirebaseAuth.instance.signOut();
+      return true;
+    } catch (e) {
+      DialogBox.showAlertBox(
+          context: context,
+          title: "Logout Failed",
+          message: "Logout From googleaccount failed",
+          onOkPressed: () {});
+      return false;
+    }
   }
 }
